pyproject.toml
[tool.poetry]
name = "latch-o11y"
version = "0.1.3"
description = "Observability for latch python backend services"
authors = ["Max Smolin <max@latch.bio>"]
license = "CC0 1.0"
readme = "README.md"
packages = [{include = "latch_o11y"}]

[tool.poetry.dependencies]
python = "^3.11"
latch-config = "^0.1.4"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.15.0"
orjson = "^3.8.5"
structlog = "^22.3.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
isort = "^5.11.4"
rich = "^13.2.0"
ruff = "^0.0.227"

[tool.black]
preview = true
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
pydocstyle = { convention = "google" }
extend-select = ["W", "D", "U", "N", "C", "B", "A", "T", "Q", "M"]
extend-ignore = [
  "A002",
  "A003",

  "E402",
  "E501",
  "E741",

  "C408",
  "C901",

  "T201",

  "D415",

  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D107",
]

[tool.pyright]
reportUnknownArgumentType = "none"
reportUnknownLambdaType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"

reportUnusedImport = "warning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
